upgrade to vs 2010

+++ trunk/EFFEngine/Src/EFFBase/md5.cpp
@@ -145,7 +145,7 @@ void MD5::update(istream& stream){
 
   while (stream.good()){
     stream.read((char *)buffer, 1024); // note that return value of read is unusable.
-    len=stream.gcount();
+    len=(int)stream.gcount();
     update(buffer, len);
   }
 
@@ -166,7 +166,7 @@ void MD5::update(ifstream& stream){
 
   while (stream.good()){
     stream.read((char *)buffer, 1024); // note that return value of read is unusable.
-    len=stream.gcount();
+    len=(int)stream.gcount();
     update(buffer, len);
   }
